+++ libqaul/src/api/contacts.rs
@@ -1,10 +1,10 @@
 //! Service contact book API
 
-use super::{users::UserID, error::QResult};
+use common::{identity::UserID, error::{Error as QaulError, Result as QaulResult}};
 
 pub fn add(id: UserID, contact: UserID) {}
 
-pub fn get(id: UserID, contact: UserID) -> QResult<UserID> {
+pub fn get(id: UserID, contact: UserID) -> QaulResult<UserID> {
     unimplemented!()
 }
