+++ libqaul/src/api/users.rs
@@ -1,9 +1,6 @@
 //! Service user managment
 
-use super::error::{Error, QResult};
-
-// FIXME: Replace with HashID or whatever?
-pub type UserID = String;
+use common::{identity::UserID, error::{Error as QaulError, Result as QaulResult}};
 
 // FIXME: Replace with actual auth token
 pub type AuthID = String;
@@ -25,12 +22,12 @@ pub struct User {
 /// Fundamentally, a user doesn't require any metadata.
 /// If a user has optionally decided to provide data
 /// about themselves, use `update` to add it.
-pub fn create() -> QResult<UserID> {
+pub fn create() -> QaulResult<UserID> {
     unimplemented!()
 }
 
 /// Provides an interface to update a user in-place
-pub fn update<F>(id: UserID, cb: F) -> QResult<()>
+pub fn update<F>(id: UserID, cb: F) -> QaulResult<()>
 where
     F: Fn(&mut User),
 {
@@ -38,7 +35,7 @@ where
 }
 
 /// Get a read-only reference to user data
-pub fn get(id: UserID) -> QResult<&'static User> {
+pub fn get(id: UserID) -> QaulResult<&'static User> {
     unimplemented!()
 }
 
@@ -47,17 +44,17 @@ pub fn get(id: UserID) -> QResult<&'static User> {
 /// After passing authentication it is possible
 /// for a user to delete themselves from a node,
 /// including all of their stored data.
-pub fn delete(id: UserID) -> QResult<()> {
+pub fn delete(id: UserID) -> QaulResult<()> {
     unimplemented!()
 }
 
 /// Authenticate as a user, returning access token
-pub fn authenticate(id: UserID, pw: String) -> QResult<AuthID> {
+pub fn authenticate(id: UserID, pw: String) -> QaulResult<AuthID> {
     unimplemented!()
 }
 
 /// Hand in an access token, ending a current session
-pub fn unauth(id: UserID, auth: AuthID) -> QResult<()> {
+pub fn unauth(id: UserID, auth: AuthID) -> QaulResult<()> {
     unimplemented!()
 }
